Post-work support
支持部分文件打包
（把测试的那一堆commit合进来了）
