+++ .github/workflows/packer.yml
@@ -17,7 +17,10 @@ jobs:
     name: Build / Cache Packer
     runs-on: windows-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
+        with: 
+          fetch-depth: 1
+          sparse-checkout: src
 
       # 缓存程序。一方面，在不同job之间需要这么做；另一方面，大约可以改善运行时间？
       # actions/cache的逻辑会在job末尾缓存打包程序；如果不命中，就自行构造程序。
@@ -25,13 +28,15 @@ jobs:
         id: cache-packer
         uses: actions/cache@v3
         with:
-          key: ${{ runner.os }}-${{ hashFiles('src/Pakcer/**') }}
-          path: Packer.exe
+          key: ${{ runner.os }}-Packer-${{ hashFiles('src/Packer/**') }}
+          path: |
+            Packer.exe
+            git2-*.dll
           lookup-only: true
 
       # 构造程序
       - name: Build Packer if not cached
-        if: steps.cache-packer.outputs.cache_hit != 'true'
+        if: steps.cache-packer.outputs.cache-hit != 'true'
         run: dotnet publish .\src\Packer\Packer.csproj -o ./ -r win-x64 -p:PublishSingleFile=true
 
   build-uploader:
@@ -39,7 +44,10 @@ jobs:
     name: Build / Cache Uploader
     runs-on: windows-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
+        with: 
+          fetch-depth: 1
+          sparse-checkout: src
 
       # 缓存程序。一方面，在不同job之间需要这么做；另一方面，大约可以改善运行时间？
       # actions/cache的逻辑会在job末尾缓存打包程序；如果不命中，就自行构造程序。
@@ -47,13 +55,13 @@ jobs:
         id: cache-uploader
         uses: actions/cache@v3
         with:
-          key: ${{ runner.os }}-${{ hashFiles('src/Uploader/**') }}
+          key: ${{ runner.os }}-Uploader-${{ hashFiles('src/Uploader/**') }}
           path: Uploader.exe
           lookup-only: true
 
       # 构造程序
       - name: Build Uploader if not cached
-        if: steps.cache-uploader.outputs.cache_hit != 'true'
+        if: steps.cache-uploader.outputs.cache-hit != 'true'
         run: dotnet publish .\src\Uploader\Uploader.csproj -o ./ -r win-x64 -p:PublishSingeFile=true
 
   initialize-release:
@@ -102,8 +110,10 @@ jobs:
         id: cache-restore
         uses: actions/cache/restore@v3
         with:
-          key: ${{ runner.os }}-${{ hashFiles('source/Pakcer/**') }}
-          path: Packer.exe
+          key: ${{ runner.os }}-${{ hashFiles('source/Packer/**') }}
+          path: |
+            Packer.exe
+            git2-*.dll
           fail-on-cache-miss: true # 前一步理应构造过的。如果不命中，肯定有问题，不如直接挂掉。
 
       - name: Check changed path on ${{ matrix.version }}
@@ -117,6 +127,7 @@ jobs:
             src/**
 
       - name: Run Packer for ${{ matrix.version }}
+        # 分发包中应当包含全部内容
         run: ./Packer --version="${{ matrix.version }}"
         # 运行逻辑：内容有更改 或 手动运行
         if: steps.check-changes.outputs.changed == 'true' || github.event_name == 'workflow_dispatch'
