+++ .gitignore
@@ -15,6 +15,10 @@ coreclr.dll
 mscordaccore.dll
 /src/Packer/Properties/launchSettings.json
 
+# libgit2sharp
+
+git2-*.dll
+
 # artifact temporary files
 artifacts/
